+++ cardenete-server/src/main/java/org/cardenete/rest/CategoriaRestController.java
@@ -3,9 +3,14 @@ package org.cardenete.rest;
 import java.util.List;
 
 import org.cardenete.entity.ResponseBean;
+import org.cardenete.enums.RolesEnum;
+import org.cardenete.entity.ArticuloBean;
 import org.cardenete.entity.CategoriaBean;
 import org.cardenete.exceptions.BeanNotFoundException;
+import org.cardenete.exceptions.EmptyListException;
+import org.cardenete.exceptions.NotAuthException;
 import org.cardenete.service.generic.GenericServiceInterface;
+import org.cardenete.validations.CheckPermission;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,6 +27,9 @@ public class CategoriaRestController {
 	@Autowired
 	private GenericServiceInterface genericService;
 
+	@Autowired
+	private CheckPermission check;
+
 	@GetMapping("/categorias/{idCategoria}")
 	public CategoriaBean getCategoria(@PathVariable int idCategoria) {
 
@@ -33,38 +41,54 @@ public class CategoriaRestController {
 
 	@GetMapping("/categorias")
 	public List<CategoriaBean> getAllCategorias() {
-		return (List<CategoriaBean>) genericService.getAll(CategoriaBean.class);
+		
+		List<CategoriaBean> listaCategorias = (List<CategoriaBean>) genericService.getAll(CategoriaBean.class);
+		if(listaCategorias.size() < 1) {
+			throw new EmptyListException("Sin resultados");
+		}else {
+			return listaCategorias;
+		}
 	}
 
 	@PostMapping("/categorias")
 	public ResponseBean addCategoria(@RequestBody CategoriaBean oCategoria) {
-		return new ResponseBean(200, String.valueOf(genericService.save(oCategoria)));
+		if (check.checkRolePermissions(RolesEnum.ADMIN.roleId)) {
+			return new ResponseBean(200, String.valueOf(genericService.save(oCategoria)));
+		} else {
+			throw new NotAuthException("No tienes suficientes permisos.");
+		}
 	}
 
 	@PutMapping("/categorias")
-	public ResponseBean updateBean(@RequestBody CategoriaBean oCategoria) {
+	public ResponseBean updateCategoria(@RequestBody CategoriaBean oCategoria) {
+		if (check.checkRolePermissions(RolesEnum.ADMIN.roleId)) {
+			// throw exception if null
+			if (genericService.get(CategoriaBean.class, oCategoria.getId()) == null) {
+				throw new BeanNotFoundException("Categoria con el id: " + oCategoria.getId() + " no encontrado.");
+			}
 
-		// throw exception if null
-		if (genericService.get(CategoriaBean.class, oCategoria.getId()) == null) {
-			throw new BeanNotFoundException("Categoria con el id: " + oCategoria.getId() + " no encontrado.");
+			return new ResponseBean(200, genericService.saveOrUpdate(oCategoria));
+		} else {
+			throw new NotAuthException("No tienes suficientes permisos.");
 		}
-
-		return new ResponseBean(200, genericService.saveOrUpdate(oCategoria));
 	}
 
 	@DeleteMapping("/categorias/{idCategoria}")
-	public ResponseBean deleteCustomer(@PathVariable int idCategoria) {
+	public ResponseBean deleteCategoria(@PathVariable int idCategoria) {
+		if (check.checkRolePermissions(RolesEnum.ADMIN.roleId)) {
+			CategoriaBean oCategoria = genericService.get(CategoriaBean.class, idCategoria);
 
-		CategoriaBean oCategoria = genericService.get(CategoriaBean.class, idCategoria);
+			// throw exception if null
+			if (oCategoria == null) {
+				throw new BeanNotFoundException("Categoria con el id: " + idCategoria + " no encontrado.");
+			}
 
-		// throw exception if null
-		if (oCategoria == null) {
-			throw new BeanNotFoundException("Categoria con el id: " + idCategoria + " no encontrado.");
+			genericService.delete(oCategoria);
+
+			return new ResponseBean(200, "Categoria con id - " + idCategoria + " borrada.");
+		} else {
+			throw new NotAuthException("No tienes suficientes permisos.");
 		}
-
-		genericService.delete(oCategoria);
-
-		return new ResponseBean(200, "Categoria con id - " + idCategoria+" borrada.");
 	}
 
 }
