+++ cardenete-server/src/main/java/org/cardenete/rest/ImagenRestController.java
@@ -2,10 +2,13 @@ package org.cardenete.rest;
 
 import java.util.List;
 
-import org.cardenete.entity.ResponseBean;
 import org.cardenete.entity.ImagenBean;
+import org.cardenete.entity.ResponseBean;
 import org.cardenete.exceptions.BeanNotFoundException;
+import org.cardenete.exceptions.EmptyListException;
+import org.cardenete.exceptions.NotAuthException;
 import org.cardenete.service.generic.GenericServiceInterface;
+import org.cardenete.validations.CheckPermission;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,6 +25,9 @@ public class ImagenRestController {
 	@Autowired
 	private GenericServiceInterface genericService;
 
+	@Autowired
+	private CheckPermission check;
+
 	@GetMapping("/imagenes/{idImagen}")
 	public ImagenBean getImagen(@PathVariable int idImagen) {
 
@@ -33,38 +39,55 @@ public class ImagenRestController {
 
 	@GetMapping("/imagenes")
 	public List<ImagenBean> getAllImagenes() {
-		return (List<ImagenBean>) genericService.getAll(ImagenBean.class);
+		
+		List<ImagenBean> listaImagenes = (List<ImagenBean>) genericService.getAll(ImagenBean.class);
+		if(listaImagenes.size() < 1) {
+			throw new EmptyListException("Sin resultados");
+		}else {
+			return listaImagenes;
+		}
+		
 	}
 
 	@PostMapping("/imagenes")
 	public ResponseBean addImagen(@RequestBody ImagenBean oImagen) {
-		return new ResponseBean(200, String.valueOf(genericService.save(oImagen)));
+		// Si no estás loggeado no puedes añadir imágenes
+		if(check.checkIsLogged()) {
+			return new ResponseBean(200, String.valueOf(genericService.save(oImagen)));			
+		}else {
+			throw new NotAuthException("No tienes suficientes permisos.");
+		}
 	}
 
 	@PutMapping("/imagenes")
-	public ResponseBean updateBean(@RequestBody ImagenBean oImagen) {
+	public ResponseBean updateImagen(@RequestBody ImagenBean oImagen) {
+		if (check.checkIsLogged() && check.checkSameUserSession(oImagen.getId_usuario().getId())) {
+			// throw exception if null
+			if (genericService.get(ImagenBean.class, oImagen.getId()) == null) {
+				throw new BeanNotFoundException("Imagen con el id: " + oImagen.getId() + " no encontrado.");
+			}
 
-		// throw exception if null
-		if (genericService.get(ImagenBean.class, oImagen.getId()) == null) {
-			throw new BeanNotFoundException("Imagen con el id: " + oImagen.getId() + " no encontrado.");
+			return new ResponseBean(200, genericService.saveOrUpdate(oImagen));
+		} else {
+			throw new NotAuthException("No tienes suficientes permisos.");
 		}
-
-		return new ResponseBean(200, genericService.saveOrUpdate(oImagen));
 	}
 
 	@DeleteMapping("/imagenes/{idImagen}")
-	public ResponseBean deleteCustomer(@PathVariable int idImagen) {
-
+	public ResponseBean deleteImagen(@PathVariable int idImagen) {
 		ImagenBean oImagen = genericService.get(ImagenBean.class, idImagen);
-
 		// throw exception if null
 		if (oImagen == null) {
 			throw new BeanNotFoundException("Imagen con el id: " + idImagen + " no encontrado.");
 		}
+		if (check.checkSameUserSession(oImagen.getId_usuario().getId())) {
 
-		genericService.delete(oImagen);
+			genericService.delete(oImagen);
 
-		return new ResponseBean(200, "Imagen con id - " + idImagen+" borrada.");
+			return new ResponseBean(200, "Imagen con id - " + idImagen + " borrada.");
+		} else {
+			throw new NotAuthException("No tienes suficientes permisos.");
+		}
 	}
 
 }
